pyproject.toml
[project]
name = "pptx-to-images"
version = "0.1.0"
description = "Convert PowerPoint slides to images using aspose.slides"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "aspose-slides>=24.0",
    "cloudconvert>=2.1.0",
    "pillow>=11.3.0",
]

[project.scripts]
pptx2img = "pptx_to_images:main"

[dependency-groups]
dev = []
